* move_to_folder:
  Move folder to folder now complete. Ready to launch!
  Moving feed to folder, end-to-end. Just needs move folder to folder.
  Assetmodel for move to folder.

+++ apps/reader/models.py
@@ -352,7 +352,7 @@ class UserSubscriptionFolders(models.Model):
         self.folders = json.encode(user_sub_folders)
         self.save()
         
-    def delete_feed(self, feed_id, in_folder):
+    def delete_feed(self, feed_id, in_folder, commit_delete=True):
         def _find_feed_in_folders(old_folders, folder_name='', multiples_found=False, deleted=False):
             new_folders = []
             for k, folder in enumerate(old_folders):
@@ -381,7 +381,7 @@ class UserSubscriptionFolders(models.Model):
         self.folders = json.encode(user_sub_folders)
         self.save()
 
-        if not multiples_found and deleted:
+        if not multiples_found and deleted and commit_delete:
             try:
                 user_sub = UserSubscription.objects.get(user=self.user, feed=feed_id)
             except Feed.DoesNotExist:
@@ -396,8 +396,8 @@ class UserSubscriptionFolders(models.Model):
                 user_sub.delete()
             MUserStory.objects(user_id=self.user.pk, feed_id=feed_id).delete()
 
-    def delete_folder(self, folder_to_delete, in_folder, feed_ids_in_folder):
-        def _find_folder_in_folders(old_folders, folder_name, feeds_to_delete):
+    def delete_folder(self, folder_to_delete, in_folder, feed_ids_in_folder, commit_delete=True):
+        def _find_folder_in_folders(old_folders, folder_name, feeds_to_delete, deleted_folder=None):
             new_folders = []
             for k, folder in enumerate(old_folders):
                 if isinstance(folder, int):
@@ -408,18 +408,22 @@ class UserSubscriptionFolders(models.Model):
                     for f_k, f_v in folder.items():
                         if f_k == folder_to_delete and folder_name == in_folder:
                             logging.user(self.user, "~FBDeleting folder '~SB%s~SN' in '%s': %s" % (f_k, folder_name, folder))
+                            deleted_folder = folder
                         else:
-                            nf, feeds_to_delete = _find_folder_in_folders(f_v, f_k, feeds_to_delete)
+                            nf, feeds_to_delete, deleted_folder = _find_folder_in_folders(f_v, f_k, feeds_to_delete, deleted_folder)
                             new_folders.append({f_k: nf})
     
-            return new_folders, feeds_to_delete
+            return new_folders, feeds_to_delete, deleted_folder
             
         user_sub_folders = json.decode(self.folders)
-        user_sub_folders, feeds_to_delete = _find_folder_in_folders(user_sub_folders, '', feed_ids_in_folder)
+        user_sub_folders, feeds_to_delete, deleted_folder = _find_folder_in_folders(user_sub_folders, '', feed_ids_in_folder)
         self.folders = json.encode(user_sub_folders)
         self.save()
         
-        UserSubscription.objects.filter(user=self.user, feed__in=feeds_to_delete).delete()
+        if commit_delete:
+          UserSubscription.objects.filter(user=self.user, feed__in=feeds_to_delete).delete()
+          
+        return deleted_folder
         
     def rename_folder(self, folder_to_rename, new_folder_name, in_folder):
         def _find_folder_in_folders(old_folders, folder_name):
@@ -442,6 +446,27 @@ class UserSubscriptionFolders(models.Model):
         user_sub_folders = _find_folder_in_folders(user_sub_folders, '')
         self.folders = json.encode(user_sub_folders)
         self.save()
+        
+    def move_feed_to_folder(self, feed_id, in_folder=None, to_folder=None):
+        user_sub_folders = json.decode(self.folders)
+        self.delete_feed(feed_id, in_folder, commit_delete=False)
+        user_sub_folders = json.decode(self.folders)
+        user_sub_folders = add_object_to_folder(int(feed_id), to_folder, user_sub_folders)
+        self.folders = json.encode(user_sub_folders)
+        self.save()
+        
+        return self
+
+    def move_folder_to_folder(self, folder_name, in_folder=None, to_folder=None):
+        user_sub_folders = json.decode(self.folders)
+        deleted_folder = self.delete_folder(folder_name, in_folder, [], commit_delete=False)
+        user_sub_folders = json.decode(self.folders)
+        user_sub_folders = add_object_to_folder(deleted_folder, to_folder, user_sub_folders)
+        self.folders = json.encode(user_sub_folders)
+        self.save()
+        
+        return self
+        
 
 class Feature(models.Model):
     """
